orvim-backend/requirements.txt
fastapi==0.111.0
pydantic==2.8.1
pydantic-settings==2.3.4
uvicorn==0.30.1
boto3==1.35.20
PyJWT==2.8.0
passlib==1.7.4
email_validator==2.2.0
SQLAlchemy==2.0.31
psycopg2==2.9.9
aio_pika==9.4.3
